+++ semcore/skeleton/src/charts/Histogram/Histogram.type.ts
@@ -10,7 +10,7 @@ declare namespace NSSkeletonHistogramChart {
     type Component = Intergalactic.Component<'svg', Props>;
 }
 
-/** @deprecated It will be removed in v18. */
+/** @deprecated It will be removed in v19. */
 export type HistogramChartSkeletonProps = NSSkeletonHistogramChart.Props;
 
 export type { NSSkeletonHistogramChart };