+++ semcore/time-picker/src/component/PickerFormat/PickerFormat.type.ts
@@ -1,9 +1,9 @@
-import type { BoxProps } from '@semcore/base-components';
+import type { NSBox } from '@semcore/base-components';
 import type { useI18n } from '@semcore/core/lib/utils/enhances/WithI18n';
 
 import type { TimePickerMeridiem } from '../TimePicker/TimePicker.type';
 
-export type TimePickerFormatProps = BoxProps & {};
+export type TimePickerFormatProps = NSBox.Props & {};
 
 export type TimePickerFormatPropsInternal = {
   /** @Internal */