+++ tools/theme/src/utils.ts
@@ -139,6 +139,14 @@ export function processTokens(config: Theme, prefix: string): ProcessedTokens {
         traverse({ node: config.baseTokens[key], path: [], prefix: `${prefix}-duration`, groupKey: 'baseTokens' });
         break;
       }
+      case 'easings': {
+        traverse({ node: config.baseTokens[key], path: [], prefix: `${prefix}-easings`, groupKey: 'baseTokens' });
+        break;
+      }
+      case 'assets': {
+        traverse({ node: config.baseTokens[key], path: [], prefix: `${prefix}-assets`, groupKey: 'baseTokens' });
+        break;
+      }
       default: {
         const k: never = key;
         throw new Error(`Please, handle key ${k}`);